Fix #101 Add Key property indicator to each parental node
Augmented the definitions for a single property within each parental node used thus far with a "Key: true" indicator that the property in question is the field to which child records must be mapped during data loading.

Added an extra property to be flagged as the Key property where necessary, i.e. within:
diagnosis
enrollment
visit

Also fixed minor typos in  a couple of property descriptions