Part 1 of  some general maintenance and tidy-up
Cleaned up guidance statements pertaining to the most appropriate use of terms  within certain controlled vocabularies.

On a node by node basis, began re-aligning the order in which properties are listed within the props of props file with the order in which the same properties are listed in strictly logical order within the nodes and props file.

Did some arbitrary tidying up, with more to come.